requirements.txt
# requirements.txt
psycopg2-binary==2.9.9
netifaces==0.11.0
scapy==2.5.0
python-dotenv==1.0.0
